+++ Bash/Installer Scripts/GitHub Projects/build-jemalloc.sh
@@ -5,7 +5,7 @@
 ##  Updated: 12.03.23
 ##  Script version: 1.0
 
-if [ "${EUID}" -eq '0' ]; then
+if [ "$EUID" -eq 0 ]; then
     echo "You must run this script without root or sudo."
     exit 1
 fi
@@ -14,27 +14,31 @@ fi
 
 script_ver=1.0
 jemalloc_ver=5.3.0
-archive_dir=jemalloc-${jemalloc_ver}
-archive_url=https://github.com/jemalloc/jemalloc/archive/refs/tags/"${jemalloc_ver}".tar.gz
+archive_dir="jemalloc-$jemalloc_ver"
+archive_url="https://github.com/jemalloc/jemalloc/archive/refs/tags/$jemalloc_ver.tar.gz"
 archive_ext="${archive_url//*.}"
 archive_name="$archive_dir.tar.${archive_ext}"
-cwd="$PWD"/jemalloc-build-script
+cwd="$PWD/jemalloc-build-script"
 install_dir=/usr/local
-pc_type=$(gcc -dumpmachine)
-user_agent='Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/121.0.0.0 Safari/537.36'
-web_repo=https://github.com/slyfox1186/script-repo
+
+CC="gcc"
+CXX="g++"
+CFLAGS="-O2 -pipe -mtune=native"
+CXXFLAGS="$CFLAGS"
+PATH="/usr/lib/ccache:$PATH"
+PKG_CONFIG_PATH="/usr/local/lib/pkgconfig:/usr/local/lib64/pkgconfig:/usr/local/share/pkgconfig:/usr/lib/pkgconfig:/usr/lib64/pkgconfig:/usr/share/pkgconfig"
+PKG_CONFIG_PATH+=":/usr/local/cuda/lib64/pkgconfig:/usr/local/cuda/lib/pkgconfig:/opt/cuda/lib64/pkgconfig:/opt/cuda/lib/pkgconfig"
+PKG_CONFIG_PATH+=":/usr/lib/x86_64-linux-gnu/pkgconfig:/usr/lib/i386-linux-gnu/pkgconfig:/usr/lib/arm-linux-gnueabihf/pkgconfig:/usr/lib/aarch64-linux-gnu/pkgconfig"
+export CC CXX CFLAGS CPPFLAGS CXXFLAGS LDFLAGS PATH PKG_CONFIG_PATH
 
 # Create output directory jemalloc
-
-if [ -d "$cwd/$archive_dir" ]; then
-    sudo rm -fr "$cwd/$archive_dir"
-fi
+[[ -d "$cwd/$archive_dir" ]] && sudo rm -fr "$cwd/$archive_dir"
 mkdir -p "$cwd/$archive_dir/build"
 
 # Download the archive file jemalloc
 
-if [ ! -f "$cwd/$archive_dir.tar.gz" ]; then
-    curl -A "$user_agent" -Lso "$cwd/$archive_dir.tar.gz" "${archive_url}"
+if [[ ! -f "$cwd/$archive_dir.tar.gz" ]]; then
+    curl -Lso "$cwd/$archive_dir.tar.gz" "${archive_url}"
 fi
 
 # Extract the archive file jemalloc
@@ -45,23 +49,23 @@ fi
 
 # Install jemalloc
 
-printf "\n%s\n%s\n\n"                        \
-    "Installing Jemalloc - v${jemalloc_ver}" \
+printf "\n%s\n%s\n\n" \
+    "Installing Jemalloc - v$jemalloc_ver" \
     '==============================================='
 
 cd "$cwd/$archive_dir" || exit 1
 ./autogen.sh
 cd build || exit 1
-../configure --prefix="$install_dir"   \
-             --disable-debug             \
-             --disable-doc               \
-             --disable-fill              \
+../configure --prefix="$install_dir" \
+             --disable-debug \
+             --disable-doc \
+             --disable-fill \
              --disable-initial-exec-tls  \
-             --disable-log               \
-             --disable-prof              \
-             --disable-stats             \
-             --enable-autogen            \
-             --enable-static             \
+             --disable-log \
+             --disable-prof \
+             --disable-stats \
+             --enable-autogen \
+             --enable-static \
              --enable-xmalloc
 echo
 if ! make "-j$(nproc --all)"; then
